requirements-dev.txt
# Dependências para desenvolvimento - versão completa com reconhecimento facial

# FastAPI e dependências core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic-settings==2.0.3

# Reconhecimento facial - OPCIONAL (comentar se causar problemas)
# Instale via: pip install face-recognition opencv-python-headless numpy
face-recognition==1.3.0
opencv-python-headless==4.8.1.78
numpy>=1.21.0,<1.25.0
Pillow==10.1.0

# Base de dados e cache
psycopg2-binary==2.9.8
redis==5.0.1
sqlalchemy==2.0.23

# Segurança e validação
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Monitoramento e logs
loguru==0.7.2
prometheus-client==0.19.0

# Utilitários
aiofiles==23.2.1
httpx==0.25.2

# Desenvolvimento
pytest==7.4.3
pytest-asyncio==0.21.1